requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
boltons @ file:///C:/ci_311/boltons_1677729932371/work
brotlipy==0.7.0
certifi @ file:///C:/b/abs_4a0polqwty/croot/certifi_1683875377622/work/certifi
cffi @ file:///C:/ci_311/cffi_1676423759166/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///C:/ci_311/colorama_1676422310965/work
conda @ file:///C:/b/abs_b5q92oum8h/croot/conda_1689272261421/work
conda-content-trust @ file:///C:/ci_311/conda-content-trust_1676467587370/work
conda-libmamba-solver @ file:///C:/b/abs_5a9xvt2a3k/croot/conda-libmamba-solver_1685032355900/work/src
conda-package-handling @ file:///C:/b/abs_ce4_vcfd0y/croot/conda-package-handling_1685024800103/work
conda_package_streaming @ file:///C:/b/abs_88a7k_wmm1/croot/conda-package-streaming_1685019697115/work
cryptography @ file:///C:/b/abs_b28tzldafl/croot/cryptography_1686613103834/work
dataclasses-json==0.5.14
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
greenlet==2.0.2
huggingface-hub==0.16.4
idna @ file:///C:/ci_311/idna_1676424932545/work
Jinja2==3.1.2
jsonpatch @ file:///tmp/build/80754af9/jsonpatch_1615747632069/work
jsonpointer==2.1
langchain==0.0.266
langsmith==0.0.23
libmambapy @ file:///C:/b/abs_f0tnqlkcbn/croot/mamba-split_1685993165748/work/libmambapy
MarkupSafe==2.1.3
marshmallow==3.20.1
menuinst @ file:///C:/ci_311/menuinst_1678730372782/work
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numexpr==2.8.5
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging @ file:///C:/b/abs_ed_kb9w6g4/croot/packaging_1678965418855/work
Pillow==10.0.0
pluggy @ file:///C:/ci_311/pluggy_1676422178143/work
pycosat @ file:///C:/ci_311/pycosat_1676438455539/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.12
pyOpenSSL @ file:///C:/b/abs_de215ipd18/croot/pyopenssl_1678965319166/work
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests @ file:///C:/b/abs_41owkd5ymz/croot/requests_1682607524657/work
ruamel.yaml @ file:///C:/ci_311/ruamel.yaml_1676439214109/work
safetensors==0.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy==2.0.20
sympy==1.12
tenacity==8.2.3
tokenizers==0.13.3
toolz @ file:///C:/ci_311/toolz_1676431406517/work
torch==2.0.1
tqdm @ file:///C:/b/abs_f76j9hg7pv/croot/tqdm_1679561871187/work
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
yarl==1.9.2
zstandard==0.19.0
